Revert "Merge branch 'Placing'"
This reverts commit 21982e81daf9acf03d4e96eeea8b9f89100c5d07, reversing
changes made to 094c29e555789c4572a8ce8f2b1a374180d03b10.

+++ Game Farm/Assets/Scripts/Player.cs
@@ -6,7 +6,7 @@ using UnityEngine.SceneManagement;
 
 public class Player : MonoBehaviour
 {
-    public double money;
+    public int money;
     public int exp = 0;
     public int lvl = 100;
     public Text MoneyTxt;
